+++ 최규화/Mod_Client/프레임워크 3.4/mainFrameWork.cpp
@@ -471,7 +471,7 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 				XMFLOAT3 mAxis = XMFloat4to3(Float4Cross(sV1, sV2));
 				mAxis = Float3Normalize(mAxis);
 				//보정을 시킨다.
-				AmendObject(mAxis, (float)stheta, obj);
+				AmendObject(mAxis, stheta, obj);
 
 				//그리고 재귀 시킨다. 왜냐하면 보정이되었으면 allpoint,tempcollisionpoint,contactpoint , penetration 모두 다 바뀌어야 하기 때문이다.
 				//재귀 후 아마 2가지 경우의수가 있다. 충돌이 일어나거나, 아니면 살짝 떠있거나.. 어쨌든 잘 해결 된다.
@@ -484,7 +484,7 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 				XMFLOAT3 mAxis = XMFloat4to3(Float4Cross(V1, V2));
 				mAxis = Float3Normalize(mAxis);
 				//보정을 시킨다.
-				AmendObject(mAxis, (float)theta, obj);
+				AmendObject(mAxis, theta, obj);
 
 				//그리고 재귀 시킨다. 왜냐하면 보정이되었으면 allpoint,tempcollisionpoint,contactpoint , penetration 모두 다 바뀌어야 하기 때문이다.
 				//재귀 후 아마 2가지 경우의수가 있다. 충돌이 일어나거나, 아니면 살짝 떠있거나.. 어쨌든 잘 해결 된다.
@@ -509,9 +509,9 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 
 
 
-					Jm.x *= obj->rb->GetMass()*(impurse + (float)ratioImpurse);
-					Jm.y *= obj->rb->GetMass()*(impurse + (float)ratioImpurse);
-					Jm.z *= obj->rb->GetMass()*(impurse + (float)ratioImpurse);
+					Jm.x *= obj->rb->GetMass()*(impurse + ratioImpurse);
+					Jm.y *= obj->rb->GetMass()*(impurse + ratioImpurse);
+					Jm.z *= obj->rb->GetMass()*(impurse + ratioImpurse);
 
 
 
@@ -520,7 +520,7 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 					auto W = obj->rb->GetAngularVelocity();
 					XMVECTOR rxi = XMLoadFloat3(&XMFloat4to3(Float4Add(fp.Pos, obj->CenterPos, false)));
 					rxi = XMVector3Cross(rxi, XMLoadFloat3(&Normal));
-					rxi *= (float)(ratioImpurse);
+					rxi *= (ratioImpurse);
 					rxi = XMVector3Transform(rxi, XMLoadFloat4x4(&obj->rb->GetIMoment()));
 
 					XMFLOAT3 ia;
@@ -659,7 +659,7 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 					XMFLOAT3 mAxis = XMFloat4to3(Float4Cross(V2, V3));
 					mAxis = Float3Normalize(mAxis);
 					//보정을 시킨다.
-					AmendObject(mAxis, (float)theta, obj);
+					AmendObject(mAxis, theta, obj);
 
 					//그리고 재귀 시킨다. 왜냐하면 보정이되었으면 allpoint,tempcollisionpoint,contactpoint , penetration 모두 다 바뀌어야 하기 때문이다.
 					//재귀 후 아마 2가지 경우의수가 있다. 충돌이 일어나거나, 아니면 살짝 떠있거나.. 어쨌든 잘 해결 된다.
@@ -702,7 +702,7 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 						XMFLOAT3 mAxis = XMFloat4to3(Float4Cross(V2, V3));
 						mAxis = Float3Normalize(mAxis);
 						//보정을 시킨다.
-						AmendObject(mAxis, (float)theta, obj);
+						AmendObject(mAxis, theta, obj);
 
 						//그리고 재귀 시킨다. 왜냐하면 보정이되었으면 allpoint,tempcollisionpoint,contactpoint , penetration 모두 다 바뀌어야 하기 때문이다.
 						//재귀 후 아마 2가지 경우의수가 있다. 충돌이 일어나거나, 아니면 살짝 떠있거나.. 어쨌든 잘 해결 된다.
@@ -726,9 +726,9 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 
 
 
-						Jm.x *= obj->rb->GetMass()*(impurse + (float)ratioImpurse);
-						Jm.y *= obj->rb->GetMass()*(impurse + (float)ratioImpurse);
-						Jm.z *= obj->rb->GetMass()*(impurse + (float)ratioImpurse);
+						Jm.x *= obj->rb->GetMass()*(impurse + ratioImpurse);
+						Jm.y *= obj->rb->GetMass()*(impurse + ratioImpurse);
+						Jm.z *= obj->rb->GetMass()*(impurse + ratioImpurse);
 
 
 
@@ -737,7 +737,7 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 						auto W = obj->rb->GetAngularVelocity();
 						XMVECTOR rxi = XMLoadFloat3(&XMFloat4to3(Float4Add(fp.Pos, obj->CenterPos, false)));
 						rxi = XMVector3Cross(rxi, XMLoadFloat3(&Normal));
-						rxi *= (float)(ratioImpurse);
+						rxi *= (ratioImpurse);
 						rxi = XMVector3Transform(rxi, XMLoadFloat4x4(&obj->rb->GetIMoment()));
 
 						XMFLOAT3 ia;
@@ -807,7 +807,7 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 
 
 					//보정을 시킨다.
-					AmendObject(mAxis, (float)theta, obj);
+					AmendObject(mAxis, theta, obj);
 					obj->rb->SetAngularVelocity(0, 0, 0);
 
 				}
@@ -835,7 +835,7 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 
 			//땅에 닿았으니 현재 속도의 y는 반감되어야 한다. 원래는 탄성계수가 있지만.. 그냥 절반 감소시킨후 부호를 -로 하자.
 			auto d = obj->rb->GetVelocity();
-			d.y = -0.01f * d.y;
+			d.y = -0.01 * d.y;
 			obj->rb->SetVelocity(d);
 			obj->rb->SetAccel(0, 0, 0);
 
@@ -853,7 +853,7 @@ void MainFrameWork::RigidBodyCollisionPlane(XMFLOAT3 & Normal, float distance, C
 
 			//땅에 닿았으니 현재 속도의 y는 반감되어야 한다. 원래는 탄성계수가 있지만.. 그냥 절반 감소시킨후 부호를 -로 하자.
 			auto d = obj->rb->GetVelocity();
-			d.y = -0.01f * d.y;
+			d.y = -0.01 * d.y;
 			obj->rb->SetVelocity(d);
 			obj->rb->SetAccel(0, 0, 0);
